Fix Threshold not being set to double? in NotificationTriggerData used when deserializing triggers running on non-English versions of PRTG (Fix #230)

+++ src/PrtgAPI/Objects/Models/NotificationTriggerData.cs
@@ -104,7 +104,7 @@ namespace PrtgAPI
 
         [DataMember(Name = "threshold")]
         [PropertyParameter(TriggerProperty.Threshold)]
-        int? Threshold { get; set; }
+        double? Threshold { get; set; }
 
         [DataMember(Name = "threshold_input")]
         string ThresholdInput { get; set; }
